feat: implement database integration with SQLAlchemy models, repositories, and services for wallet management

+++ app/schemas.py
@@ -1,13 +1,14 @@
+from decimal import Decimal
 from pydantic import BaseModel, Field, field_validator
 
 
 class OperationRequest(BaseModel):
     wallet_name: str = Field(..., max_length=127)
-    amount: float
+    amount: Decimal
     description: str | None = Field(None, max_length=255)
 
     @field_validator('amount')
-    def amount_must_be_positive(cls, v: float) -> float:
+    def amount_must_be_positive(cls, v: Decimal) -> Decimal:
         if v <= 0:
             raise ValueError("Amount must be positive")
         return v
@@ -22,10 +23,10 @@ class OperationRequest(BaseModel):
 
 class CreateWalletRequest(BaseModel):
     name: str = Field(..., max_length=127)
-    initial_balance: float = 0
+    initial_balance: Decimal = 0
 
     @field_validator('initial_balance')
-    def balance_not_negative(cls, v: float) -> float:
+    def balance_not_negative(cls, v: Decimal) -> Decimal:
         if v < 0:
             raise ValueError("Initial balance cannot be negative")
         return v
